overlay/00-coreos-static.conf: fixate more users/groups
We have a bunch of users and groups added at compose time which are not
currently fixated and so risk drifting in the future as packages get
added and removed.

Fixate those.

Ideally we should have CI checks for this even so that we keep this list
up to date.

The longer-term fix for this is to reduce as much as possible the number
of sysusers truly required at compose time which e.g. requires packaging
changes.