Konflux: generate rpm-lockfile-prototype input file
The resulting config file will be fed to `rpm-lockfile-protype` which
in turns will generate the hermeto config.
This will allow us to enable hermetic builds in Konflux.

Run with :
```
 ./ci/generate_rpm_lockfile_config manifest.yaml
 ````

This creates `rpms.in.yaml`.

The intent is to trigger this in the `bump-lockfile` job of jenkins.

Requires https://github.com/konflux-ci/rpm-lockfile-prototype/pull/107
And https://github.com/konflux-ci/rpm-lockfile-prototype/pull/104