build-rootfs: add overrides/rootfs API
Similarly to the previous commit, this supports the longstanding cosa
API of having an `overrides/rootfs` dir with bits that get overlaid on
top of the final rootfs.

Except here the dir is in the build context directory. cosa can use this
API to wire its `overrides/rootfs` to keep the same API, but `podman
build` users can also use this directly.

This is almost equivalent to having a
`COPY overrides/rootfs/ /target-rootfs` instruction after the
`build-rootfs` but before rechunking.

The nice thing about folding it into build-rootfs is that it can more
cleanly be made conditional than a Containerfile directive, and we can
support the override being fed as a bind-mount (as cosa will use it)
instead of forcing it to be in the context dir.