Re-activate next-devel branch
This syncs over settings and lockfiles from testing-devel. We are going
to do a F42 build of next to do the container signing migration [1] and
so we'll need `next` and `testing` to differ for now. We were about to
do this anyway because `next-devel` and `next` will be going to F43
in the next round anyway.