konflux: update pipeline reference
This is a follow-up of [1]. With this update, the images will start
containing `org.opencontainers.image.{source,version}` as label and
annotations. This partially enables tagging the image with the version,
see [2] for more details.
This update also contains [3] which fixes the issue we are currently
hitting in Konflux.
Finally, this pipeline ref is using pinned task as reference, which
is required to comply to Conforma `redhat` policy. More details in [4].

[1] https://github.com/coreos/fedora-coreos-config/pull/3732
[2] https://gitlab.com/fedora/infrastructure/konflux/tenants-config/-/merge_requests/72
[3] https://github.com/konflux-ci/build-definitions/pull/2785
[4] https://gitlab.com/fedora/bootc/tekton-catalog/-/issues/5