build-rootfs: mutate-os-release in a postprocess
This converts the mutate-os-release code into a postprocess. This
is needed because some other postprocess scripts read os-release
and perform conditional logic based on the version.

This commit also removes the now no longer needed 99dracut-coreos-version
dracut module and alos the horrible hack added in 2baab77
to rm systemd-gpt-auto-generator in the initramfs. Now that our
postprocess scripts happen properly as part of the rpm-ostree
compose and not after, we no longer need these hacks.