versionary: pull Y from the arch-matching lockfile
Versionary pull the Y component of the version ID from the lockfile
timestamp (if it exists).
Instead of always pulling from the x86 manifest, open the one lockfile
matching the machine architecture.
This will fix a corner case of the `bump-lockfile` job failing when
creating the lockfile for the first time: on the remote multi-arch
builders, the x86 lockfile does not exist yet
(it will be there in the next commit.)