tests/manual: add bisect ability for test run to be INCONCLUSIVE
Sometimes a particular build in the builds.json can't be tested.
Maybe it's a multi-arch build and the build for that version didn't
succeed, or maybe we're querying information populated by the release
job and this version was never released. Either way let's add an
option for an INCONCLUSIVE state.

An example of running this is at [1]

[1] https://github.com/coreos/fedora-coreos-tracker/issues/2066#issuecomment-3548064906