drop fedora-bootc submodule
We're now completely moved over to building via container [1], which
means that we are no longer using the git submodule for manifest
definitions for bootc but rather the manifests delivered with the
build container. Thus we are always "deriving: true" and we can delete
any config that was for the "deriving: false" condition.

[1] https://github.com/coreos/fedora-coreos-tracker/issues/1969